server_py/requirements.txt
amqp==2.6.1
appnope==0.1.0
asgiref==3.2.10
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.2
celery-redbeat==0.13.0
certifi==2020.11.8
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
Django==3.1.8
django-admin-autocomplete-filter==0.6.1
django-ckeditor==6.0.0
django-tables2==2.3.3
djangorestframework==3.12.2
elasticsearch==7.9.1
html5lib==1.1
hyperlink==21.0.0
icalendar==4.0.7
idna==2.10
incremental==21.3.0
iso8601==0.1.13
itemadapter==0.2.0
itemloaders==1.0.4
iteration-utilities==0.11.0
jmespath==0.10.0
kombu==4.6.11
lxml==4.6.3
mechanize==0.4.5
mock==4.0.3
parsel==1.6.0
Protego==0.1.16
psycopg2cffi==2.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.4
queuelib==1.5.0
redis==3.5.3
requests==2.25.1
rtyaml==1.0.0
scrapelib==1.2.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.2
sqlparse==0.3.1
tenacity==7.0.0
Twisted==21.2.0
urllib3==1.26.4
uWSGI==2.0.19.1
vine==1.3.0
w3lib==1.22.0
waitress==1.4.4
webencodings==0.5.1
xmltodict==0.12.0
zope.interface==5.2.0
